ejercicios_practica/requirements.txt
Django==3.2.2
requests==2.25.1
django-filter==2.4.0
djangorestframework==3.12.4
django-rest-auth==0.9.5
# Swagger:
uritemplate==3.0.1
pyyaml==5.4.1
# Drivers para trabajar con Postgres en linux:
psycopg2-binary>=2.8